requirements.txt
streamlit>=1.28.0
yfinance>=0.2.28
pandas>=2.0.0
plotly>=5.17.0
pandas-datareader==0.10.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0